refactor: split inline test modules from 5 final userspace-dp files
Third and final batch of the pure-code-motion test-split modularity work
(after #4779 poll_stages, #4780's eight-file batch, and #4781's nine-file
batch). Relocate the tail `#[cfg(test)] mod tests { ... }` block out of the
five remaining >150-test-line modules into sibling `<name>_tests.rs`,
loaded via `#[cfg(test)] #[path = "..."] mod tests;`. Part of the
#4404/#4408/#4421 modularity backlog; this exhausts the clean-tail-block
inline-test-module cohort (all remaining candidates are under ~150 test
lines).

  xsk_ffi.rs                  (174 test lines out)
  afxdp/shared_umem.rs        (167 out)
  afxdp/wg/tai64n.rs          (166 out)
  fairness.rs                 (165 out)
  afxdp/coordinator/status.rs (150 out)

Pure code motion, verified per file:
- Relocated body is byte-identical: verbatim inner-body move (not dedent),
  then rustfmt'd (edition 2024, syntax-aware) purely to canonicalise indent.
- Each parent's production code before the test module is byte-identical to
  origin/master; the only parent change is the three-line
  `mod tests {...}` -> `#[path] mod tests;` swap (parents not reformatted).
- Test count preserved per file: 7/7/8/22/5 #[test]/#[tokio::test] before
  and after. `use super::*` retained; the `#[path]` child keeps its module
  parent, so `super::` paths unchanged.

Validation: cargo build clean (all five #[path] splits resolve); FULL
cargo test --release to follow before merge. No production behavior change.
Refs #4404 #4408 #4421.
